+++ services/ui-src/src/forms/mcpar/mcpar.json
@@ -36,7 +36,7 @@
               {
                 "id": "contactName",
                 "type": "text",
-                "validation": "textOptional",
+                "validation": "text",
                 "props": {
                   "label": "A.2a Contact name",
                   "hint": "First and last name of the contact person. <br/> States that do not wish to list a specific individual on the report are encouraged to use a department or program-wide email address that will allow anyone with questions to quickly reach someone who can provide answers."
@@ -262,6 +262,7 @@
                 "type": "checkbox",
                 "validation": "checkbox",
                 "props": {
+                  "styleAsOptional": true,
                   "label": "B.III.1 Data validation entity",
                   "hint": "Select the state agency/division or contractor tasked with evaluating the validity of encounter data submitted by MCPs.</br>Encounter data validation includes verifying the accuracy, completeness, timeliness, and/or consistency of encounter data records submitted to the state by Medicaid managed care plans. Validation steps may include pre-acceptance edits and post-acceptance analyses. See Glossary in Excel Workbook for more information.",
                   "choices": [
