+++ services/ui-src/src/utils/other/parsing.test.tsx
@@ -1,7 +1,10 @@
-import { render, screen } from "@testing-library/react";
+import { render, within, screen } from "@testing-library/react";
 import DOMPurify from "dompurify";
 // utils
-import { parseCustomHtml } from "./parsing";
+import { parseCustomHtml, useUser } from "utils";
+import { mockReportContext, mockStateUser } from "utils/testing/setupJest";
+import { ReportContext, TextField } from "components";
+import { useFormContext } from "react-hook-form";
 
 jest.mock("dompurify", () => ({
   sanitize: jest.fn((el) => el),
@@ -36,6 +39,40 @@ const testElementArray = [
   },
 ];
 
+const mockTrigger = jest.fn();
+const mockRhfMethods = {
+  register: () => {},
+  setValue: () => {},
+  getValues: jest.fn(),
+  trigger: mockTrigger,
+};
+const mockUseFormContext = useFormContext as unknown as jest.Mock<
+  typeof useFormContext
+>;
+jest.mock("react-hook-form", () => ({
+  useFormContext: jest.fn(() => mockRhfMethods),
+}));
+const mockGetValues = (returnValue: any) =>
+  mockUseFormContext.mockImplementation((): any => ({
+    ...mockRhfMethods,
+    getValues: jest.fn().mockReturnValue(returnValue),
+  }));
+
+jest.mock("utils/auth/useUser");
+const mockedUseUser = useUser as jest.MockedFunction<typeof useUser>;
+
+const textFieldComponent = (
+  <ReportContext.Provider value={mockReportContext}>
+    <TextField
+      name="testTextField"
+      label="test-label"
+      placeholder="test-placeholder"
+      data-testid="test-text-field"
+      styleAsOptional
+    />
+  </ReportContext.Provider>
+);
+
 const testComponent = <div>{parseCustomHtml(testElementArray)}</div>;
 
 describe("Test parseCustomHtml", () => {
@@ -56,5 +93,16 @@ describe("Test parseCustomHtml", () => {
 
   test("Type 'html' is sanitized and parsed", () => {
     expect(sanitizationSpy).toHaveBeenCalled();
+  });
+});
+
+describe("Test labelTextWithOptional", () => {
+  test("labelTextWithOptional parses a label with added html", () => {
+    mockedUseUser.mockReturnValue(mockStateUser);
+    mockGetValues(undefined);
+    const { getByText } = render(textFieldComponent);
+    const label = getByText("test-label");
+    const optional = within(label).getByText("(optional)");
+    expect(optional).toBeVisible();
   });
 });
