--- services/ui-src/src/utils/other/styleAsOptional.ts
@@ -1,4 +1,0 @@
-import { parseCustomHtml } from "./parsing";
-
-export const labelTextWithOptional = (label: any) =>
-  parseCustomHtml(label + "<span class='optional-text'> (optional)</span>");
